Create an offset to trigger touch right/left events
Add a gap on x axis to trigger the touch right/left events, so there
is a margin for the event to happen. This prevents a 'tremble effect',
where, when moving an object via these events listeners, it seems like
it's shaking. This offset makes room so the object can be stationary.